Add files via upload
A case study on loan applicants. This is just data wrangling/cleansing exercise on the entire dataset using EDA techniques to identify:

If the applicant is likely to repay the loan, then not approving the loan results in a loss of business to the company
If the applicant is not likely to repay the loan, i.e. he/she is likely to default, then approving the loan may lead to a financial loss for the company.